# swagger.yaml
openapi: 3.0.0
info:
  title: API Quản lý Nông Trại Nấm
  version: 1.0.0
  description: Tài liệu API cho Hệ thống Quản lý Nuôi trồng Nấm
servers:
  - url: /
    description: Máy chủ đang phục vụ Swagger (tự dùng host và cổng hiện tại)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Pagination:
      type: "object"
      properties:
        totalItems:
          type: "integer"
        currentPage:
          type: "integer"
          minimum: 1
        totalPages:
          type: "integer"
        pageSize:
          type: "integer"
          minimum: 1
...
    FinancialSummary:
      type: "object"
      properties:
        currency:
          type: "string"
          enum: ["VND"]
        totalHarvestKg:
          type: "number"
        totalSoldKg:
          type: "number"
        revenue:
          type: "string"
        totalCost:
          type: "string"
        profit:
          type: "string"
        costsByCategory:
          type: "object"
          properties:
            MATERIAL:
              type: "string"
            TOOL:
              type: "string"
            FERTILIZER:
              type: "string"
            OTHER:
              type: "string"
        profitMarginPercent:
          type: "string"
          nullable: true
        costPerHarvestKg:
          type: "string"
          nullable: true
    FinancialSeriesPoint:
      type: "object"
      properties:
        period:
          type: "string"
        revenue:
          type: "string"
        totalCost:
          type: "string"
        profit:
          type: "string"
    FinancialReportRow:
      type: "object"
      properties:
        batchId:
          type: "integer"
          minimum: 1
        batchCode:
          type: "string"
        status:
          type: "string"
        facility:
          type: "string"
        mushroom:
          type: "string"
        currency:
          type: "string"
          enum: ["VND"]
        totalHarvestKg:
          type: "number"
        totalSoldKg:
          type: "number"
        revenue:
          type: "string"
        totalCost:
          type: "string"
        profit:
          type: "string"
        costsByCategory:
          type: "object"
          properties:
            MATERIAL:
              type: "string"
            TOOL:
              type: "string"
            FERTILIZER:
              type: "string"
            OTHER:
              type: "string"
        profitMarginPercent:
          type: "string"
          nullable: true
        costPerHarvestKg:
          type: "string"
          nullable: true
        series:
          type: "array"
          items:
            $ref: "#/components/schemas/FinancialSeriesPoint"
...
  parameters:
    ReportFrom:
      name: from
      in: query
      schema: { type: string, format: date }
      description: Ngày bắt đầu, mặc định là đầu năm hiện tại
    ReportTo:
      name: to
      in: query
      schema: { type: string, format: date }
      description: Ngày kết thúc, mặc định là hôm nay
    FacilityId:
      name: facilityId
      in: query
      schema: { type: integer }
    MushroomId:
      name: mushroomId
      in: query
      schema: { type: integer }
    GroupBy:
      name: groupBy
      in: query
      schema: { type: string, enum: [day, month], default: month }
security:
  - bearerAuth: []
paths:
...
  /api/reports/overview:
    get:
      summary: Xem thống kê tổng hợp hệ thống
      tags: [Báo cáo]
      parameters:
        - $ref: '#/components/parameters/ReportFrom'
        - $ref: '#/components/parameters/ReportTo'
        - $ref: '#/components/parameters/FacilityId'
        - $ref: '#/components/parameters/MushroomId'
        - $ref: '#/components/parameters/GroupBy'
      responses:
        '200': { description: Thống kê tổng hợp }
...
  /api/reports/{type}/export:
    get:
      summary: Xuất báo cáo CSV, XLSX hoặc PDF
      tags: [Báo cáo]
      parameters:
        - name: type
          in: path
          required: true
          schema: { type: string, enum: [overview, cultivation, classifier, audit, financial] }
        - name: format
          in: query
          required: true
          schema: { type: string, enum: [csv, xlsx, pdf] }
      responses:
        '200':
          description: File báo cáo
          content:
            text/csv: {}
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet: {}
            application/pdf: {}
        '422': { description: Báo cáo quá lớn, cần thu hẹp bộ lọc }
...
  /api/reports/financial:
    get:
      summary: "Báo cáo tài chính theo ngày bán/ngày chi (manager/admin)"
      tags: ["Tài chính"]
      parameters: [{"$ref":"#/components/parameters/ReportFrom"},{"$ref":"#/components/parameters/ReportTo"},{"$ref":"#/components/parameters/FacilityId"},{"$ref":"#/components/parameters/MushroomId"},{"$ref":"#/components/parameters/GroupBy"},{"name":"status","in":"query","schema":{"type":"string","enum":["PREPARATION","INCUBATION","FRUITING","HARVESTING","COMPLETED","FAILED"]}},{"name":"page","in":"query","schema":{"type":"integer","minimum":1,"default":1}},{"name":"limit","in":"query","schema":{"type":"integer","minimum":1,"default":20,"maximum":100}}]
      responses:
        "200":
          description: "Thành công"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    type: "array"
                    items:
                      $ref: "#/components/schemas/FinancialReportRow"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                  summary:
                    $ref: "#/components/schemas/FinancialSummary"
                  series:
                    type: "array"
                    items:
                      $ref: "#/components/schemas/FinancialSeriesPoint"
                  period:
                    type: "object"
                    properties:
                      from:
                        type: "string"
                        format: "date-time"
                      to:
                        type: "string"
                        format: "date-time"
                      groupBy:
                        type: "string"
                        enum: ["day","month"]
        "400":
          description: "Dữ liệu không hợp lệ"
        "401":
          description: "Token thiếu hoặc không hợp lệ"
        "403":
          description: "Không đủ quyền"
        "404":
          description: "Không tìm thấy dữ liệu"
        "409":
          description: "Dữ liệu/trạng thái thay đổi; tải lại trước khi thử lại"
        "500":
          description: "Lỗi máy chủ"
...